rename : feature -> features
