+++ src/widgets/diary-visibility-controls/ui/DiaryVisibilityControls.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Container } from './DiaryVisibilityControls.styled';
-import { VisibilityButton } from '../../../feature/diary-write';
+import { VisibilityButton } from '../../../features/diary-write';
 import { DiaryVisibilityControlsProps } from '../model/type';
 
 export const DiaryVisibilityControls = ({
